requirements.txt
tensorflow_gpu==1.15
scipy
pandas
sklearn
